fix(clippy): resolve 34 clippy warnings across lib and test code
Fixes: type_complexity in patterns.rs (type alias), Duration readability
in pane_tiers/recorder_replay/runtime/config_profiles, clone_on_copy in
5 proptest files, manual_range_contains in 5 files, unused import, unused
mut, len_zero, int_plus_one, absurd_extreme_comparisons, needless_borrow,
manual_string_new, unnecessary_cast, suboptimal_flops, many_single_char_names,
needless_range_loop.

+++ crates/frankenterm-core/tests/proptest_resize_invariants.rs
@@ -895,7 +895,7 @@ proptest! {
         telemetry.absorb(&report);
         prop_assert!(telemetry.total_passes >= before_passes + n_transitions as u64,
             "absorb should add report passes to telemetry");
-        prop_assert_eq!(telemetry.total_checks, report.total_checks() as u64);
+        prop_assert_eq!(telemetry.total_checks, report.total_checks());
     }
 }
